+++ src/bin/keycloakify/buildOptions/index.ts
@@ -1,1 +1,1 @@
-export * from "./buildOptions2";
+export * from "./buildOptions";
